debugged the finance_calculator.py script

+++ python_introduction/finance_calculator.py
@@ -1,11 +1,11 @@
 # A script to calculate the projected savings with interest of a user
 
 # Variables to receive user's monthly income and expenses respectively
-monthly_income = int(input("Enter your monthly income: "))
-monthly_expenses = int(input("Enter your monthly expenses: "))
+monthly_income = float(input("Enter your monthly income: "))
+monthly_expenses = float(input("Enter your monthly expenses: "))
 
 # Calculating the monthly savings
-monthly_savings = int(monthly_income - monthly_expenses)
+monthly_savings = float(monthly_income - monthly_expenses)
 
 # calculating the projected savings and interest rate after 1 year
 projected_savings = float(monthly_savings * 12 + (monthly_savings * 12 * 0.05))
